% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_pro_files.R
\name{convert_doy_to_datetime}
\alias{convert_doy_to_datetime}
\title{Convert day of year decimal time to datetime}
\usage{
convert_doy_to_datetime(doy_decimal, year, timezone = "GMT")
}
\arguments{
\item{doy_decimal}{Numeric vector. Decimal day of year values.}

\item{year}{Numeric. The year for the conversion.}

\item{timezone}{Character. Timezone for the conversion. Default is "GMT".}
}
\value{
POSIXct datetime vector.
}
\description{
Converts a decimal day of year format (e.g., 123.5 = day 123,
12:00) to a proper datetime object.
}
\examples{
\dontrun{
# Convert day 100.5 (April 10th, 12:00) of 2024 to datetime
convert_doy_to_datetime(100.5, 2024, "GMT")
}

}
\keyword{internal}
