% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_cruise_id.R
\name{extract_flowcam_columns}
\alias{extract_flowcam_columns}
\title{Parse the cruise ID column of an EcoTaxa file of type FlowCam.}
\usage{
extract_flowcam_columns(
  ecotaxa_file,
  pattern = flowcam_patterns,
  debug = FALSE
)
}
\arguments{
\item{ecotaxa_file}{(character) The unquoted name of a tibble or data frame
in the R environment that reflecte the data in the EcoTaxa file.}

\item{pattern}{(list) A list of FlowCam patterns to match.}

\item{debug}{(logical) A logical value indicating whether to enable debug
mode. Defaults to \code{FALSE}.}
}
\value{
A data frame or tibble of the input EcoTaxa data with additional
columns reflecting the parsed values of \code{column_id}.
}
\description{
Parses the \code{cruise_id} column of an EcoTaxa file of
type FlowCam.
}
\note{
A helper function to \code{parse_cruise_id}
}
