% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_cruise_id.R
\name{parse_cruise_id}
\alias{parse_cruise_id}
\title{Parse the cruise ID column of an EcoTaxa file}
\usage{
parse_cruise_id(ecotaxa_file, debug = FALSE)
}
\arguments{
\item{ecotaxa_file}{(character) The unquoted name of a tibble or data frame
in the R environment that reflecte the data in the EcoTaxa file.}

\item{debug}{A logical value indicating whether to enable debug mode.
Defaults to \code{FALSE}.}
}
\value{
A list with two elements:
\item{extracted_file}{A data frame with extracted columns.}
\item{pattern}{A string indicating the matched pattern type ("flowcam",
"moc", "uvp", or "none").}
}
\description{
\code{parse_cruise_id} Parses the \code{cruise_id} column of an
EcoTaxa file of type MOC, UVP, or FlowCam.
}
