% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure_parallel.R
\name{setup_parallel_ecotaxa}
\alias{setup_parallel_ecotaxa}
\title{Set up parallel processing for ecotaxaLoadR functions}
\usage{
setup_parallel_ecotaxa(workers = future::availableCores() - 1)
}
\arguments{
\item{workers}{Number of parallel workers. Defaults to available cores - 1.}
}
\description{
Set up parallel processing for ecotaxaLoadR functions
}
