% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_eco_taxa.R
\name{load_eco_taxa}
\alias{load_eco_taxa}
\title{Load and minimally process downloaded EcoTaxa data}
\usage{
load_eco_taxa(file_path, debug = FALSE)
}
\arguments{
\item{file_path}{(character) Quoted path to the EcoTaxa TSV file.}
}
\value{
A data frame with the processed EcoTaxa data, including additional
columns for cruise, moc, net, fraction, and select object dimensions
converted from pixels to millimeters.
}
\description{
This function loads and processes EcoTaxa data from a TSV file.
It ensures that the data is properly formatted and annotated for further
analysis.
}
\examples{
\dontrun{
eco_taxa_df <- load_eco_taxa("path/to/eco_taxa_file.tsv")
print(eco_taxa_df)
}

}
